+++ apps/front/.svelte-kit/output/server/nodes/1.js
@@ -1,8 +1,0 @@
-
-
-export const index = 1;
-let component_cache;
-export const component = async () => component_cache ??= (await import('../entries/fallbacks/error.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/1.Djp_K7L4.js","_app/immutable/chunks/DsnmJJEf.js","_app/immutable/chunks/BSZVTExI.js","_app/immutable/chunks/BOBn-JNp.js","_app/immutable/chunks/VUbJ2AoM.js","_app/immutable/chunks/CiwgAdwy.js","_app/immutable/chunks/BQfWZw2H.js"];
-export const stylesheets = [];
-export const fonts = [];