+++ apps/front/.svelte-kit/output/server/nodes/2.js
@@ -1,8 +1,0 @@
-
-
-export const index = 2;
-let component_cache;
-export const component = async () => component_cache ??= (await import('../entries/pages/_page.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/2.DxJOVwMm.js","_app/immutable/chunks/DsnmJJEf.js","_app/immutable/chunks/BSZVTExI.js","_app/immutable/chunks/BOBn-JNp.js","_app/immutable/chunks/VUbJ2AoM.js","_app/immutable/chunks/D88A_OO5.js","_app/immutable/chunks/CWV3WZP_.js","_app/immutable/chunks/BqmXiWVO.js","_app/immutable/chunks/BauaJNm8.js","_app/immutable/chunks/CiwgAdwy.js","_app/immutable/chunks/BQfWZw2H.js"];
-export const stylesheets = [];
-export const fonts = [];