+++ apps/front/.svelte-kit/output/server/nodes/3.js
@@ -1,8 +1,0 @@
-
-
-export const index = 3;
-let component_cache;
-export const component = async () => component_cache ??= (await import('../entries/pages/cards/_page.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/3.BcygrSQH.js","_app/immutable/chunks/DsnmJJEf.js","_app/immutable/chunks/BSZVTExI.js","_app/immutable/chunks/BOBn-JNp.js","_app/immutable/chunks/ZInn0ksZ.js","_app/immutable/chunks/CBsudzpE.js","_app/immutable/chunks/BQfWZw2H.js","_app/immutable/chunks/BNYWS3Kj.js","_app/immutable/chunks/BauaJNm8.js","_app/immutable/chunks/CWV3WZP_.js","_app/immutable/chunks/DYI_lUPm.js","_app/immutable/chunks/D-V-_hmm.js","_app/immutable/chunks/DhnkYgoe.js","_app/immutable/chunks/DmDSLnt0.js"];
-export const stylesheets = [];
-export const fonts = [];